# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetClass:
    enum:
    - us_equity
    - crypto
    type: string
    x-enum-varnames:
    - AlpacaAssetClassUSEquity
    - AlpacaAssetClassCrypto
  github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetResponse:
    properties:
      class:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetClass'
      easy_to_borrow:
        type: boolean
      exchange:
        type: string
      fractionable:
        type: boolean
      id:
        type: string
      marginable:
        type: boolean
      min_order_size:
        type: number
      min_trade_increment:
        type: number
      name:
        type: string
      price_increment:
        type: number
      shortable:
        type: boolean
      status:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetStatus'
      symbol:
        type: string
      tradable:
        type: boolean
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - AlpacaAssetStatusActive
    - AlpacaAssetStatusInactive
...
  internal_api_handlers.AssetsResponse:
    properties:
      assets:
        items:
          $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  internal_api_handlers.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@stackservice.com
    name: API Support
    url: http://www.stackservice.com/support
  description: GenZ Web3 Multi-Chain Investment Platform API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stack Service API
  version: "1.0"
paths:
...
  /api/v1/assets:
    get:
      description: Retrieve a list of all tradable assets (stocks, ETFs) with optional
        filtering
      parameters:
      - default: active
        description: Asset status filter (active, inactive)
        in: query
        name: status
        type: string
      - description: Asset class filter (us_equity, crypto)
        in: query
        name: asset_class
        type: string
      - description: Exchange filter (NASDAQ, NYSE, ARCA, BATS)
        in: query
        name: exchange
        type: string
      - default: true
        description: Filter by tradability
        in: query
        name: tradable
        type: boolean
      - description: Filter by fractional shares support
        in: query
        name: fractionable
        type: boolean
      - description: Filter by short selling support
        in: query
        name: shortable
        type: boolean
      - description: Filter by easy-to-borrow status
        in: query
        name: easy_to_borrow
        type: boolean
      - description: Search by symbol or name
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 100
        description: Items per page (max 500)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.AssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Get all tradable assets
      tags:
      - assets
  /api/v1/assets/{symbol_or_id}:
    get:
      description: Retrieve detailed information about a specific asset by symbol
        or asset ID
      parameters:
      - description: Asset symbol (e.g., AAPL) or Asset ID (UUID)
        in: path
        name: symbol_or_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.AlpacaAssetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Get asset details
      tags:
      - assets
  /api/v1/assets/exchange/{exchange}:
    get:
      description: Retrieve assets listed on a specific exchange
      parameters:
      - description: Exchange code (NASDAQ, NYSE, ARCA, BATS)
        in: path
        name: exchange
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.AssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Get assets by exchange
      tags:
      - assets
  /api/v1/assets/popular:
    get:
      description: Retrieve a curated list of popular stocks and ETFs for quick access
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.AssetsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Get popular assets
      tags:
      - assets
  /api/v1/assets/search:
    get:
      description: Search for assets by symbol or company name
      parameters:
      - description: Search query (symbol or name)
        in: query
        name: q
        required: true
        type: string
      - default: 20
        description: Maximum results to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api_handlers.AssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.ErrorResponse'
      summary: Search assets
      tags:
      - assets
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"